requirements 2.txt
# Mega Brain - Python Dependencies
# Hooks in .claude/hooks/ use only stdlib + PyYAML
# Scripts in scripts/ may require additional packages (dev-only, not distributed)
PyYAML>=6.0
requests>=2.31.0
python-dotenv>=1.0.0
